Add DirtyReadTransaction as a strategy type.
This inherits from the Transaction strategy, but
sets the isolation type as "read uncommitted".
This allows nested transactions, and transactions
in different threads, to read data that hasn't been
committed yet.

This is helpful for controller specs, when using
Factory Girl for fixtures. DatabaseCleaner starts
a transaction, creates the fixture data, but then
a controller tries to access that data in a
separate transaction, and the data hasn't been
committed yet, so it looks like the fixtures
aren't working.